Extract processor steps logic into ProcessorStepsService
Move canHavePreviousStep, canReplaceStep, getProcessorStepsProperties,
DISABLED_SIBLING_STEPS, and related step-property constants out of
CamelComponentSchemaService into a new dedicated ProcessorStepsService.
Consolidate the disabled-sibling and special-processor constants in
special-processors.constants.ts and update all callers accordingly.
