+++ packages/ui/src/components/Visualization/Custom/hooks/duplicate-step.hook.tsx
@@ -8,7 +8,7 @@ import { CatalogModalContext } from '../../../../dynamic-catalog/catalog-modal.p
 import { SourceSchemaType } from '../../../../models/camel/source-schema-type';
 import { EntityType } from '../../../../models/entities';
 import { AddStepMode, IVisualizationNode } from '../../../../models/visualization/base-visual-entity';
-import { CamelComponentSchemaService } from '../../../../models/visualization/flows/support/camel-component-schema.service';
+import { ProcessorStepsService } from '../../../../models/visualization/flows/support/processor-steps.service';
 import { EntitiesContext } from '../../../../providers/entities.provider';
 import { VisibleFlowsContext } from '../../../../providers/visible-flows.provider';
 import { updateIds } from '../../../../utils/update-ids';
@@ -108,7 +108,7 @@ export const useDuplicateStep = (vizNode: IVisualizationNode) => {
 
       // Set an empty model to clear the graph, Fixes an issue rendering child nodes incorrectly
       if (parentVizNode?.getNodeInteraction().canHaveSpecialChildren) {
-        const stepsProperties = CamelComponentSchemaService.getProcessorStepsProperties(
+        const stepsProperties = ProcessorStepsService.getProcessorStepsProperties(
           parentVizNode.data.primaryNodeId?.name as keyof ProcessorDefinition,
         );
         if (
